Show available budget when approving a funding case
The available budget is the budget of the funding program minus the sum
of all approved amounts of funding cases with the funding program. It is
only displayed if a budget is defined. (It is only informational, i.e.
an amount greater than the available budget can still be approved.)