Allow funding case types to specify funding case status to which new application processes can be added to
Previously the funding case status to add new application processes was
hard coded to `open`. Now it can be specified in the properties of the
`FundingCaseType` entity with the key `applicationAddableStatusList`.
The new default is now to always create a new funding case. (This change
only affects non-combined funding case types.)