Allow to optionally specify custom tags for application editor, application form, and review sidebar
Previously funding case types where forced to provide an AngularJS
module with a custom application editor directive. This is optionally
now. A default application editor is used if not specified in the
funding case type. Also the application form as well as the review
sidebar can be exchanged by custom ones. (See changes in documentation.)