Add a contact relation type `Relationship`
This contact relation type matches a contact that has a relationship.
The relationship type may be limited to various types. The type of the
related contacts may be limited to various types. The related contacts
may be limited to those in any of the given groups.

The contact relation type `ContactTypeRelationship` is replaced by the
new type.